Separated score and solve
